app/requirements.txt
Authlib==1.2.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.4
idna==3.4
pycparser==2.21
requests==2.31.0
urllib3==2.0.5
validators==0.22.0
weaviate-client==3.24.1
